Switch to ghcup for installation method
The current installation method is dependent on debian ghc
and cabal packaging, which often is slow to be updated.
Additionally, ghcup is becoming a standard installation technique
so we probably don't want to reinvent the wheel here.

It also provides an easier path to support windows and arm based
images.