Delete old dockerfiles
There isn't much point to keeping them around, they out of date and
not the newest pattern. If someone wanted a dockerfile with older versions
they would be better off copying the pattern from the new dockerfiles.

It is convenient to do this now as hadolint has been added to check
the ARM related changes, but these old ones trigger failures.